feat(sprint1-b): config, domain models, repos, migrations, and seed script
B-1: internal/config/config.go - env-based config with dev defaults
B-2: 5 domain models (activity, enrollment, order, behavior, notification)
     - all uint64 IDs aligned with DDL BIGINT
     - GORM tags matching DDL column specs
B-3: AutoMigrate list prepared (5 models, don't modify main.go)
B-4: migrations/001~007 {up,down}.sql - DDL reference for all 7 tables
B-5: 4 repository files (activity, enrollment, order, behavior)
     - full GORM implementations, not stubs
B-6: scripts/seed/main.go - 5 users + 20 activities

Refactored:
- internal/domain/user.go: uint -> uint64 for DDL alignment
- internal/repository/user_repository.go: FindByID id type -> uint64

+++ backend/internal/domain/user.go
@@ -6,7 +6,7 @@ import (
 )
 
 type User struct {
-	ID           uint           `gorm:"primaryKey" json:"id"`
+	ID           uint64         `gorm:"primaryKey" json:"id"`
 	Phone        string         `gorm:"uniqueIndex;not null;size:20" json:"phone"`
 	Username     string         `gorm:"not null;size:50" json:"username"`
 	PasswordHash string         `gorm:"not null" json:"-"`
